+++ src/event.ts
@@ -39,7 +39,7 @@ class TypeId {
 }
 
 /**
- * Emits an event to all listeners
+ * Listen for an event or category of events
  *
  * @param target - a vidar object
  * @param type - the id of the type (can contain subtypes, such as
@@ -54,6 +54,26 @@ export function subscribe (target: VidarObject, type: string, listener: <T exten
   listeners.get(target).push(
     { type: new TypeId(type), listener }
   )
+}
+
+/**
+ * Remove an event listener
+ *
+ * @param target - a vidar object
+ * @param type - the id of the type (can contain subtypes, such as
+ * "type.subtype")
+ * @param listener
+ */
+export function unsubscribe (target: VidarObject, listener: <T extends Event>(T) => void): void {
+  // Make sure `listener` has been added with `subscribe`.
+  if (!listeners.has(target) ||
+  !listeners.get(target).map(pair => pair.listener).includes(listener)) {
+    throw new Error('No matching event listener to remove')
+  }
+
+  const removed = listeners.get(target)
+    .filter(pair => pair.listener !== listener)
+  listeners.set(target, removed)
 }
 
 /**